+++ src/com/mtautumn/edgequest/window/layers/MouseSelection.java
@@ -8,9 +8,9 @@ public class MouseSelection {
 	public static void draw(Renderer r) {
 		Color.white.bind();
 		
-		int posX = getMousePosX(r);
-		int posY = getMousePosY(r);
-		int blockSize = r.dataManager.settings.blockSize;
+		float posX = getMousePosX(r);
+		float posY = getMousePosY(r);
+		float blockSize = r.dataManager.settings.blockSize;
 		
 		if (r.dataManager.system.isMouseFar) drawFarSelection(r, posX, posY, blockSize);
 		else drawNearSelection(r, posX, posY, blockSize);
@@ -19,13 +19,13 @@ public class MouseSelection {
 	}
 	
 	
-	private static int getMousePosX(Renderer r) {
+	private static float getMousePosX(Renderer r) {
 		double coordsOffsetX = offsetX(r);
-		return (int)((r.dataManager.system.mouseX - coordsOffsetX)*r.dataManager.settings.blockSize);
+		return (float)((r.dataManager.system.mouseX - coordsOffsetX)*r.dataManager.settings.blockSize);
 	}
-	private static int getMousePosY(Renderer r) {
+	private static float getMousePosY(Renderer r) {
 		double coordsOffsetY = offsetY(r);
-		return (int)((r.dataManager.system.mouseY - coordsOffsetY)*r.dataManager.settings.blockSize);
+		return (float)((r.dataManager.system.mouseY - coordsOffsetY)*r.dataManager.settings.blockSize);
 	}
 	
 	
@@ -36,16 +36,16 @@ public class MouseSelection {
 		return r.dataManager.system.screenY - Double.valueOf(r.dataManager.settings.screenHeight) / Double.valueOf(2 * r.dataManager.settings.blockSize);
 	}
 	
-	private static void drawFarSelection(Renderer r, int posX, int posY, int blockSize) {
+	private static void drawFarSelection(Renderer r, float posX, float posY, float blockSize) {
 		r.drawTexture(r.textureManager.getTexture("selectFar"), posX, posY, blockSize, blockSize);
 
 	}
-	private static void drawNearSelection(Renderer r, int posX, int posY, int blockSize) {
+	private static void drawNearSelection(Renderer r, float posX, float posY, float blockSize) {
 		r.drawTexture(r.textureManager.getTexture("select"), posX, posY, blockSize, blockSize);
 
 	}
-	private static void drawFlag(Renderer r, int posX, int posY, int blockSize) {
-		r.drawTexture(r.textureManager.getTexture("selectFlag"), posX, posY - (int)(0.4375 * blockSize), (int)(blockSize * 1.25), (int)(blockSize*1.4375));
+	private static void drawFlag(Renderer r, float posX, float posY, float blockSize) {
+		r.drawTexture(r.textureManager.getTexture("selectFlag"), posX, posY - (0.4375f * blockSize), blockSize * 1.25f, blockSize*1.4375f);
 	}
 
 }
